return null for number and percentage if input is empty
To be compliant with the angular input[number], the ui-number-mask
and ui-percentage-mask should parse an empty value to null and not
to undefined (or an empty string). This has an advantage, that performing
calculations with null is possible (e.g. 5 + null === 5). If
undefined or a string would be returned, then the model
has an illegal value and an explicit check for undefined is required
in all calculations...